Invalidate http cache when a package has been deleted from a feed
This change adds support for invalidating the http cache on a per package id basis when a 404 is received. As part of that finding the original casing of the package has been moved into the GetDependency request where the package is downloaded to disk.

Perf improvements along with this:
* Packages are downloaded only if they are the best match, previously this was done a per feed basis which caused unneeded downloads
* Added additional caching for nuspecs and dependency info

Fixes https://github.com/NuGet/Home/issues/5023
Fixes https://github.com/NuGet/Home/issues/5012